Fixed #32 -- Changed nose to pytest.
