+++ tests/tasks.py
@@ -1,11 +1,12 @@
 import json
-import redis
 import time
+
+import redis
 
 from tasktiger import RetryException
 from tasktiger.retry import fixed
 
-from .config import *
+from .config import DELAY, TEST_DB
 from .utils import get_tiger
 
 tiger = get_tiger()
@@ -14,16 +15,20 @@ tiger = get_tiger()
 def simple_task():
     pass
 
+
 @tiger.task()
 def decorated_task(*args, **kwargs):
     pass
 
+
 def exception_task():
     raise Exception('this failed')
+
 
 @tiger.task(queue='other')
 def task_on_other_queue():
     pass
+
 
 def file_args_task(filename, *args, **kwargs):
     open(filename, 'w').write(json.dumps({
@@ -31,18 +36,22 @@ def file_args_task(filename, *args, **kwargs):
         'kwargs': kwargs,
     }))
 
+
 @tiger.task(hard_timeout=DELAY)
 def long_task_killed():
-    time.sleep(DELAY*2)
+    time.sleep(DELAY * 2)
 
-@tiger.task(hard_timeout=DELAY*2)
+
+@tiger.task(hard_timeout=DELAY * 2)
 def long_task_ok():
     time.sleep(DELAY)
+
 
 @tiger.task(unique=True)
 def unique_task(value=None):
     conn = redis.Redis(db=TEST_DB, decode_responses=True)
     conn.lpush('unique_task', value)
+
 
 @tiger.task(lock=True)
 def locked_task(key, other=None):
@@ -53,8 +62,9 @@ def locked_task(key, other=None):
     time.sleep(DELAY)
     conn.delete(key)
 
+
 @tiger.task(queue='batch', batch=True)
-def batch_task(params): 
+def batch_task(params):
     conn = redis.Redis(db=TEST_DB, decode_responses=True)
     try:
         conn.rpush('batch_task', json.dumps(params))
@@ -63,6 +73,7 @@ def batch_task(params):
     if any(p['args'][0] == 10 for p in params if p['args']):
         raise Exception('exception')
 
+
 @tiger.task(queue='batch')
 def non_batch_task(arg):
     conn = redis.Redis(db=TEST_DB, decode_responses=True)
@@ -70,23 +81,26 @@ def non_batch_task(arg):
     if arg == 10:
         raise Exception('exception')
 
+
 def retry_task():
     raise RetryException()
+
 
 def retry_task_2():
     raise RetryException(method=fixed(DELAY, 1),
                          log_error=False)
 
+
 def verify_current_task():
     conn = redis.Redis(db=TEST_DB, decode_responses=True)
 
     try:
-        tasks = tiger.current_tasks
+        tiger.current_tasks
     except RuntimeError:
         # This is expected (we need to use current_task)
-
         task = tiger.current_task
         conn.set('task_id', task.id)
+
 
 @tiger.task(batch=True, queue='batch')
 def verify_current_tasks(tasks):
@@ -99,6 +113,7 @@ def verify_current_tasks(tasks):
 
         tasks = tiger.current_tasks
         conn.rpush('task_ids', *[t.id for t in tasks])
+
 
 @tiger.task()
 def sleep_task(delay=10):
