+++ tests/test_redis_scripts.py
@@ -1,6 +1,7 @@
 import unittest
 
-from .utils import *
+from .utils import get_tiger
+
 
 class RedisScriptsTestCase(unittest.TestCase):
     def setUp(self):
@@ -23,85 +24,85 @@ class RedisScriptsTestCase(unittest.TestCase):
 
     def test_zadd_nx(self):
         entries = self._test_zadd('nx')
-        self.assertEqual(entries, [('key2', 1.0), ('key1', 2.0)])
+        assert entries == [('key2', 1.0), ('key1', 2.0)]
 
     def test_zadd_xx(self):
         entries = self._test_zadd('xx')
-        self.assertEqual(entries, [('key1', 3.0)])
+        assert entries == [('key1', 3.0)]
 
     def test_zadd_min(self):
         entries = self._test_zadd('min')
-        self.assertEqual(entries, [('key2', 0.0), ('key1', 2.0)])
+        assert entries == [('key2', 0.0), ('key1', 2.0)]
 
     def test_zadd_max(self):
         entries = self._test_zadd('max')
-        self.assertEqual(entries, [('key2', 2.0), ('key1', 4.0)])
+        assert entries == [('key2', 2.0), ('key1', 4.0)]
 
     def test_zpoppush_1(self):
         self.conn.zadd('src', a=1, b=2, c=3, d=4)
         result = self.scripts.zpoppush('src', 'dst', 3, None, 10)
-        self.assertEqual(result, ['a', 'b', 'c'])
+        assert result == ['a', 'b', 'c']
 
         src = self.conn.zrange('src', 0, -1, withscores=True)
-        self.assertEqual(src, [('d', 4.0)])
+        assert src == [('d', 4.0)]
 
         dst = self.conn.zrange('dst', 0, -1, withscores=True)
-        self.assertEqual(dst, [('a', 10.0), ('b', 10.0), ('c', 10.0)])
+        assert dst == [('a', 10.0), ('b', 10.0), ('c', 10.0)]
 
     def test_zpoppush_2(self):
         self.conn.zadd('src', a=1, b=2, c=3, d=4)
         result = self.scripts.zpoppush('src', 'dst', 100, None, 10)
-        self.assertEqual(result, ['a', 'b', 'c', 'd'])
+        assert result == ['a', 'b', 'c', 'd']
 
         src = self.conn.zrange('src', 0, -1, withscores=True)
-        self.assertEqual(src, [])
+        assert src == []
 
         dst = self.conn.zrange('dst', 0, -1, withscores=True)
-        self.assertEqual(dst, [('a', 10.0), ('b', 10.0), ('c', 10.0), ('d', 10.0)])
+        assert dst == [('a', 10.0), ('b', 10.0), ('c', 10.0), ('d', 10.0)]
 
     def test_zpoppush_3(self):
         self.conn.zadd('src', a=1, b=2, c=3, d=4)
         result = self.scripts.zpoppush('src', 'dst', 3, 2, 10)
-        self.assertEqual(result, ['a', 'b'])
+        assert result == ['a', 'b']
 
         src = self.conn.zrange('src', 0, -1, withscores=True)
-        self.assertEqual(src, [('c', 3.0), ('d', 4.0)])
+        assert src == [('c', 3.0), ('d', 4.0)]
 
         dst = self.conn.zrange('dst', 0, -1, withscores=True)
-        self.assertEqual(dst, [('a', 10.0), ('b', 10.0)])
+        assert dst == [('a', 10.0), ('b', 10.0)]
 
     def test_zpoppush_withscores_1(self):
         self.conn.zadd('src', a=1, b=2, c=3, d=4)
         result = self.scripts.zpoppush('src', 'dst', 3, None, 10, withscores=True)
-        self.assertEqual(result, ['a', '1', 'b', '2', 'c', '3'])
+        assert result == ['a', '1', 'b', '2', 'c', '3']
 
         src = self.conn.zrange('src', 0, -1, withscores=True)
-        self.assertEqual(src, [('d', 4.0)])
+        assert src == [('d', 4.0)]
 
         dst = self.conn.zrange('dst', 0, -1, withscores=True)
-        self.assertEqual(dst, [('a', 10.0), ('b', 10.0), ('c', 10.0)])
+        assert dst == [('a', 10.0), ('b', 10.0), ('c', 10.0)]
 
     def test_zpoppush_withscores_2(self):
         self.conn.zadd('src', a=1, b=2, c=3, d=4)
         result = self.scripts.zpoppush('src', 'dst', 100, None, 10, withscores=True)
-        self.assertEqual(result, ['a', '1', 'b', '2', 'c', '3', 'd', '4'])
+        assert result == ['a', '1', 'b', '2', 'c', '3', 'd', '4']
 
         src = self.conn.zrange('src', 0, -1, withscores=True)
-        self.assertEqual(src, [])
+        assert src == []
 
         dst = self.conn.zrange('dst', 0, -1, withscores=True)
-        self.assertEqual(dst, [('a', 10.0), ('b', 10.0), ('c', 10.0), ('d', 10.0)])
+        assert dst == [('a', 10.0), ('b', 10.0), ('c', 10.0), ('d', 10.0)]
 
     def test_zpoppush_withscores_3(self):
         self.conn.zadd('src', a=1, b=2, c=3, d=4)
         result = self.scripts.zpoppush('src', 'dst', 3, 2, 10, withscores=True)
-        self.assertEqual(result, ['a', '1', 'b', '2'])
+        assert result == ['a', '1', 'b', '2']
 
         src = self.conn.zrange('src', 0, -1, withscores=True)
-        self.assertEqual(src, [('c', 3.0), ('d', 4.0)])
+        assert src == [('c', 3.0), ('d', 4.0)]
 
         dst = self.conn.zrange('dst', 0, -1, withscores=True)
-        self.assertEqual(dst, [('a', 10.0), ('b', 10.0)])
+        assert dst == [('a', 10.0), ('b', 10.0)]
 
     def test_zpoppush_on_success_1(self, **kwargs):
         """
@@ -116,16 +117,16 @@ class RedisScriptsTestCase(unittest.TestCase):
                 count=2, score=None, new_score=10,
                 on_success=('update_sets', 'val', 'remove_set', 'add_set'),
                 **kwargs)
-        self.assertEqual(result, ['a', 'b'])
+        assert result == ['a', 'b']
 
         src = self.conn.zrange('src', 0, -1, withscores=True)
-        self.assertEqual(src, [('c', 3.0), ('d', 4.0)])
+        assert src == [('c', 3.0), ('d', 4.0)]
 
         dst = self.conn.zrange('dst', 0, -1, withscores=True)
-        self.assertEqual(dst, [('a', 10.0), ('b', 10.0)])
+        assert dst == [('a', 10.0), ('b', 10.0)]
 
-        self.assertEqual(self.conn.smembers('remove_set'), set(['val']))
-        self.assertEqual(self.conn.smembers('add_set'), set(['val']))
+        assert self.conn.smembers('remove_set') == set(['val'])
+        assert self.conn.smembers('add_set') == set(['val'])
 
     def test_zpoppush_on_success_2(self, **kwargs):
         """
@@ -137,16 +138,16 @@ class RedisScriptsTestCase(unittest.TestCase):
                 count=2, score=0, new_score=10,
                 on_success=('update_sets', 'val', 'remove_set', 'add_set'),
                 **kwargs)
-        self.assertEqual(result, [])
+        assert result == []
 
         src = self.conn.zrange('src', 0, -1, withscores=True)
-        self.assertEqual(src, [('a', 1.0), ('b', 2.0), ('c', 3.0), ('d', 4.0)])
+        assert src == [('a', 1.0), ('b', 2.0), ('c', 3.0), ('d', 4.0)]
 
         dst = self.conn.zrange('dst', 0, -1, withscores=True)
-        self.assertEqual(dst, [])
+        assert dst == []
 
-        self.assertEqual(self.conn.smembers('remove_set'), set(['val']))
-        self.assertEqual(self.conn.smembers('add_set'), set([]))
+        assert self.conn.smembers('remove_set') == set(['val'])
+        assert self.conn.smembers('add_set') == set()
 
     def test_zpoppush_on_success_3(self, **kwargs):
         """
@@ -158,16 +159,16 @@ class RedisScriptsTestCase(unittest.TestCase):
                 count=4, score=None, new_score=10,
                 on_success=('update_sets', 'val', 'remove_set', 'add_set'),
                 **kwargs)
-        self.assertEqual(result, ['a', 'b', 'c', 'd'])
+        assert result == ['a', 'b', 'c', 'd']
 
         src = self.conn.zrange('src', 0, -1, withscores=True)
-        self.assertEqual(src, [])
+        assert src == []
 
         dst = self.conn.zrange('dst', 0, -1, withscores=True)
-        self.assertEqual(dst, [('a', 10), ('b', 10), ('c', 10), ('d', 10)])
+        assert dst == [('a', 10), ('b', 10), ('c', 10), ('d', 10)]
 
-        self.assertEqual(self.conn.smembers('remove_set'), set([]))
-        self.assertEqual(self.conn.smembers('add_set'), set(['val']))
+        assert self.conn.smembers('remove_set') == set()
+        assert self.conn.smembers('add_set') == set(['val'])
 
     def test_zpoppush_ignore_if_exists_1(self):
         self.conn.zadd('src', a=1, b=2, c=3, d=4)
@@ -178,16 +179,16 @@ class RedisScriptsTestCase(unittest.TestCase):
                 count=2, score=None, new_score=10,
                 on_success=('update_sets', 'val', 'remove_set', 'add_set'),
                 if_exists=('noupdate',))
-        self.assertEqual(result, [])
+        assert result == []
 
         src = self.conn.zrange('src', 0, -1, withscores=True)
-        self.assertEqual(src, [('c', 3.0), ('d', 4.0)])
+        assert src == [('c', 3.0), ('d', 4.0)]
 
         dst = self.conn.zrange('dst', 0, -1, withscores=True)
-        self.assertEqual(dst, [('a', 5.0), ('b', 5.0)])
+        assert dst == [('a', 5.0), ('b', 5.0)]
 
-        self.assertEqual(self.conn.smembers('remove_set'), set(['val']))
-        self.assertEqual(self.conn.smembers('add_set'), set(['val']))
+        assert self.conn.smembers('remove_set') == set(['val'])
+        assert self.conn.smembers('add_set') == set(['val'])
 
     def test_zpoppush_ignore_if_exists_2(self):
         self.test_zpoppush_on_success_2(if_exists=('noupdate',))
@@ -206,20 +207,20 @@ class RedisScriptsTestCase(unittest.TestCase):
                 on_success=('update_sets', 'val', 'remove_set', 'add_set',
                             'add_set_if_exists'),
                 if_exists=('add', 'if_exists', 20, 'min'))
-        self.assertEqual(result, ['b'])
+        assert result == ['b']
 
         src = self.conn.zrange('src', 0, -1, withscores=True)
-        self.assertEqual(src, [('c', 3.0), ('d', 4.0)])
+        assert src == [('c', 3.0), ('d', 4.0)]
 
         dst = self.conn.zrange('dst', 0, -1, withscores=True)
-        self.assertEqual(dst, [('a', 5.0), ('b', 10.0)])
+        assert dst == [('a', 5.0), ('b', 10.0)]
 
         if_exists = self.conn.zrange('if_exists', 0, -1, withscores=True)
-        self.assertEqual(if_exists, [('a', expected_if_exists_score)])
+        assert if_exists == [('a', expected_if_exists_score)]
 
-        self.assertEqual(self.conn.smembers('remove_set'), set(['val']))
-        self.assertEqual(self.conn.smembers('add_set'), set(['val']))
-        self.assertEqual(self.conn.smembers('add_set_if_exists'), set(['val']))
+        assert self.conn.smembers('remove_set') == set(['val'])
+        assert self.conn.smembers('add_set') == set(['val'])
+        assert self.conn.smembers('add_set_if_exists') == set(['val'])
 
     def test_zpoppush_min_if_exists_1(self):
         self._test_zpoppush_min_if_exists(20)
@@ -235,28 +236,28 @@ class RedisScriptsTestCase(unittest.TestCase):
     def test_srem_if_not_exists_1(self):
         self.conn.sadd('set', 'member')
         result = self.scripts.srem_if_not_exists('set', 'member', 'other_key')
-        self.assertEqual(result,  1)
-        self.assertEqual(self.conn.smembers('set'), set([]))
+        assert result == 1
+        assert self.conn.smembers('set') == set()
 
     def test_srem_if_not_exists_2(self):
         self.conn.sadd('set', 'member')
         self.conn.set('other_key', 0)
         result = self.scripts.srem_if_not_exists('set', 'member', 'other_key')
-        self.assertEqual(result,  0)
-        self.assertEqual(self.conn.smembers('set'), set(['member']))
+        assert result == 0
+        assert self.conn.smembers('set') == set(['member'])
 
     def test_delete_if_not_in_zsets_1(self):
         self.conn.set('key', 0)
         self.conn.zadd('z2', other=0)
         result = self.scripts.delete_if_not_in_zsets('key', 'member',
             ['z1', 'z2'])
-        self.assertEqual(result,  1)
-        self.assertEqual(self.conn.exists('key'), 0)
+        assert result == 1
+        assert self.conn.exists('key') == 0
 
     def test_delete_if_not_in_zsets_2(self):
         self.conn.set('key', 0)
         self.conn.zadd('z2', member=0)
         result = self.scripts.delete_if_not_in_zsets('key', 'member',
             ['z1', 'z2'])
-        self.assertEqual(result,  0)
-        self.assertEqual(self.conn.exists('key'), 1)
+        assert result == 0
+        assert self.conn.exists('key') == 1
