+++ tests/utils.py
@@ -3,7 +3,8 @@ import redis
 import structlog
 from tasktiger import TaskTiger, Worker, fixed
 
-from .config import *
+from .config import DELAY, TEST_DB
+
 
 class Patch(object):
     """
@@ -26,6 +27,7 @@ class Patch(object):
     def __exit__(self, *args):
         setattr(self.orig_obj, self.func_name, self.orig_func)
 
+
 def get_tiger():
     """
     Sets up logging and returns a new tasktiger instance.
@@ -41,9 +43,9 @@ def get_tiger():
         # doing a single worker run.
         'SELECT_TIMEOUT': 0,
 
-        'ACTIVE_TASK_UPDATE_TIMEOUT': 2*DELAY,
+        'ACTIVE_TASK_UPDATE_TIMEOUT': 2 * DELAY,
 
-        'LOCK_RETRY': DELAY*2.,
+        'LOCK_RETRY': DELAY * 2.,
 
         'DEFAULT_RETRY_METHOD': fixed(DELAY, 2),
 
@@ -53,6 +55,7 @@ def get_tiger():
     })
     tiger.log.setLevel(logging.CRITICAL)
     return tiger
+
 
 def external_worker(n=None):
     """
