+++ tox.ini
@@ -2,6 +2,6 @@
 envlist = py27,py34,py35
 
 [testenv]
-commands=nosetests
-deps=nose
+commands=pytest
+deps=pytest
      psutil
